requirements.txt
asgiref==3.6.0; python_version >= '3.7'
certifi==2022.12.7; python_version >= '3.6'
charset-normalizer==3.1.0; python_version >= '3.7'
django-jazzmin==2.6.0
django==4.2
djangorestframework==3.14.0
idna==3.4; python_version >= '3.5'
pillow==9.5.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3; python_version >= '3.5'
tzdata==2023.3; sys_platform == 'win32'
urllib3==1.26.15; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'